Enhance validation and documentation in general
This is a small change that is primarily to clean up some of the issues
with the initial documentation in the README file.

However, in addition to that, and in order to make this both easier to
use with other libraries and to allow for data masking to be optional
in nature, this also adjusts the signature and handling done by the
validate() function.

Lastly, the body() wrapper has been augmented so that it is possible to
optionally have a stack trace generated into the JSON body if desired,
by setting an environment variable in the worker context.